Move `errors` into `error_handling` package.
