+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/RholangCLI.scala
@@ -13,7 +13,7 @@ import coop.rchain.metrics.{Metrics, NoopSpan, Span}
 import coop.rchain.models._
 import coop.rchain.rholang.interpreter.Runtime.RhoISpace
 import coop.rchain.rholang.interpreter.accounting._
-import coop.rchain.rholang.interpreter.errors._
+import coop.rchain.rholang.interpreter.error_handling.errors._
 import coop.rchain.rholang.interpreter.storage.StoragePrinter
 import coop.rchain.shared.Resources
 import monix.eval.{Coeval, Task}
