+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/accounting/RholangMethodsCostsSpec.scala
@@ -11,7 +11,7 @@ import coop.rchain.models.rholang.implicits._
 import coop.rchain.rholang.interpreter.Runtime.RhoISpace
 import coop.rchain.rholang.interpreter._
 import coop.rchain.rholang.interpreter.accounting.Chargeable._
-import coop.rchain.rholang.interpreter.errors.InterpreterError
+import coop.rchain.rholang.interpreter.error_handling.errors.InterpreterError
 import coop.rchain.rspace.history.Branch
 import coop.rchain.rspace._
 import coop.rchain.rspace.{RSpace, ReplayRSpace}
